start formatting course table
