+++ resources/js/pages/reports/SchedulingReport/CourseTableRow.vue
@@ -1,9 +1,13 @@
 <template>
   <tr>
     <Td class="course-column">
-      <div>{{ course.subject }} {{ course.catalogNumber }}</div>
-      <div>
-        {{ course.title }}
+      <div>{{ course.title }}</div>
+      <div class="tw-text-xs tw-text-neutral-500 tw-flex tw-flex-wrap tw-gap-1">
+        <div>{{ course.subject }} {{ course.catalogNumber }}</div>
+        •
+        <div>{{ course.courseType }}</div>
+        •
+        <div>{{ course.courseLevel }}</div>
       </div>
     </Td>
     <Td
@@ -25,6 +29,9 @@
 import { Td } from "@/components/Table";
 import LeaveChip from "@/components/LeaveChip.vue";
 import { Instructor, Term, Leave, Course, TimelessCourse } from "@/types";
+import { doesCourseNumberMatchSearchTerm } from "./doesCourseMatchSearchTerm";
+import { doesInstructorNameMatchSearchTerm } from "./doesInstructorNameMatchSearchTerm";
+
 defineProps<{
   course: TimelessCourse;
   search: string;
@@ -32,5 +39,22 @@ defineProps<{
   listOfTermInstructors: Instructor[][];
   currentTerm: Term | null;
 }>();
+
+function isFallTerm(term: Term) {
+  return term.name.includes("Fall");
+}
 </script>
-<style scoped></style>
+<style scoped>
+.term-data-column.term-data-column--current {
+  background: #fffcf0;
+  border-top: 1px solid #fde68a;
+}
+
+.term-data-column.term-data-column--current.term-data-column--fall {
+  border-left: 2px solid #fde68a;
+}
+
+.term-data-column.term-data-column--fall {
+  border-left: 2px solid #f3f3f3;
+}
+</style>
