+++ resources/js/pages/reports/SchedulingReport/doesCourseMatchSearchTerm.ts
@@ -1,8 +1,10 @@
 import { Course } from "@/types";
 
-export function doesCourseMatchSearchTerm(course: Course, searchTerm: string) {
-  const courseTitle =
-    `${course.subject} ${course.catalogNumber} ${course.classSection}`.toLowerCase();
+export function doesCourseNumberMatchSearchTerm(
+  course: Course,
+  searchTerm: string,
+) {
+  const courseTitle = `${course.subject} ${course.catalogNumber}`.toLowerCase();
 
   return courseTitle.includes(searchTerm.toLowerCase());
 }
