+++ resources/js/pages/reports/SchedulingReport/doesInstructorNameMatchSearchTerm.ts
@@ -1,0 +1,10 @@
+import type { Instructor } from "@/types";
+export function doesInstructorNameMatchSearchTerm(
+  instructor: Instructor,
+  searchTerm: string,
+) {
+  return (
+    instructor.surName.toLowerCase().includes(searchTerm.toLowerCase()) ||
+    instructor.givenName.toLowerCase().includes(searchTerm.toLowerCase())
+  );
+}
